% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_requests.r
\name{get_search_subreddit}
\alias{get_search_subreddit}
\title{Search for subreddits based on a keyword.}
\usage{
get_search_subreddit(keyword)
}
\arguments{
\item{keyword}{The keyword used for subreddit search.}

\item{limit}{The maximum number of results to retrieve (optional).}
}
\value{
A response object containing subreddit search results.
}
\description{
This function searches for subreddits on Reddit based on a specified keyword.
}
\examples{
response <- get_search_subreddit("dogs")
cat("Subreddit search response:", response, "\n")

}
