% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_requests.r
\name{get_subredit_titles}
\alias{get_subredit_titles}
\title{Get titles from a specified subreddit.}
\usage{
get_subredit_titles(subreddit)
}
\arguments{
\item{subreddit}{The name of the subreddit.}
}
\value{
A response object containing subreddit titles.
}
\description{
This function retrieves hot posts data from specific subreddit.
}
\examples{
response <- get_subredit_titles("cats")
cat("Subreddit titles response:", response, "\n")

}
