% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_threads_summary.R
\name{get_threads_summary}
\alias{get_threads_summary}
\title{Summarize the search results of Reddit threads by specific keyword}
\usage{
get_threads_summary(keyword)
}
\arguments{
\item{keyword}{A string representing what you want to search for on Reddit.}
}
\value{
A list of three plots: upvote_ratio_plot, top_subreddits_plot, top_nouns_plot
}
\description{
Summarize the search results of Reddit threads by specific keyword
}
\examples{
\dontrun{
get_threads_summary("Mazzy Star")
get_threads_summary("UBCO")
}  
}
