% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_requests.r
\name{get_token}
\alias{get_token}
\title{Get an access token for Reddit API authentication.}
\usage{
get_token(
  client_id = CLIENT_ID,
  client_secret = CLIENT_SECRET,
  user_agent = USER_AGENT
)
}
\arguments{
\item{client_id}{Reddit API client ID.}

\item{client_secret}{Reddit API client secret.}

\item{user_agent}{User agent string for API requests.}
}
\value{
An access token for Reddit API authentication.
}
\description{
This function retrieves an access token for authentication with the Reddit API
}
\examples{
token <- get_token()
cat("Access token:", token, "\n")

}
