% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subreddit_highfreq_sentiment.r
\name{subreddit_highfreq_sentiment}
\alias{subreddit_highfreq_sentiment}
\title{Analyze the high-frequency words in subreddit titles and visualize their sentiment.}
\usage{
subreddit_highfreq_sentiment(keyword, test = 0)
}
\arguments{
\item{keyword}{A keyword to search for subreddits.}
}
\value{
A ggplot2 object representing a bar graph of word frequency by sentiment.
}
\description{
This function selects a subreddit based on a keyword, retrieves titles from the subreddit,
and analyzes the high-frequency words in the titles while visualizing their sentiment.
}
\examples{
subreddit_highfreq_sentiment("cats")

}
