% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_word_freq.R
\name{user_word_freq}
\alias{user_word_freq}
\title{Generate a Word Cloud from User Content on Reddit}
\usage{
user_word_freq(username, content_type)
}
\arguments{
\item{username}{A character string specifying the Reddit username to fetch content for.}

\item{content_type}{A character string specifying the type of content to fetch from the user's Reddit profile.
Valid options are "comments" or "submitted" for user comments or submitted posts, respectively.}
}
\value{
An \code{htmlwidget} object that displays a word cloud visualization of the most frequent nouns and
adjectives found in the user's content. Returns \code{NULL} if there are errors in fetching or processing the content.
}
\description{
This function fetches content from Reddit based on the specified username and content type,
preprocesses the text to filter nouns and adjectives, and generates a word cloud visualization.
}
\examples{
\dontrun{
 user_word_freq("Techno_superbowl", "comments")
}
}
