% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_threads_summary.R
\name{get_threads_summary}
\alias{get_threads_summary}
\title{Summarize Reddit Thread Search Results}
\usage{
get_threads_summary(keyword)
}
\arguments{
\item{keyword}{A character string representing the keyword to search for on Reddit.}
}
\value{
A list containing three plots:
\describe{
\item{upvote_ratio_plot}{A histogram showing the distribution of upvote ratios.}
\item{top_subreddits_plot}{A bar plot displaying the top subreddits by number of threads.}
\item{top_nouns_plot}{A bar plot showing the top 20 frequent nouns in thread titles and texts.}
}
}
\description{
This function retrieves and summarizes search results of Reddit threads by a specific keyword.
}
\examples{
\dontrun{
  get_threads_summary("Mazzy Star")
  get_threads_summary("UBCO")
}
}
\references{
The function uses Reddit's API for data retrieval.

The function utilizes the udpipe R package for natural language processing.
}
