% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_subreddit.r
\name{select_subreddit}
\alias{select_subreddit}
\title{Select and display the top 5 subreddits related to a keyword.}
\usage{
select_subreddit(keyword, test = 0)
}
\arguments{
\item{keyword}{The keyword used for subreddit search.}
}
\value{
The name of the selected subreddit based on user input.
}
\description{
This function performs a search for subreddits related to a given keyword and
displays the top 5 subreddits with the highest number of subscribers, then
prompts the user to select one of them.
}
\examples{
subreddit <- select_subreddit("cats")

}
